rename: useNotification -> notificationStore
